blocking workspace project invites from incorrect resolver
