+++ packages/server/modules/workspaces/domain/types.ts
@@ -1,20 +1,1 @@
-import { InviteResourceTarget } from '@/modules/serverinvites/domain/types'
-import { WorkspaceInviteResourceType } from '@/modules/workspaces/domain/constants'
-import { WorkspaceRoles } from '@speckle/shared'
-
-declare module '@/modules/serverinvites/domain/types' {
-  interface InviteResourceTargetTypeMap {
-    workspace: 'workspace'
-  }
-}
-
-declare module '@/modules/serverinvites/helpers/core' {
-  interface ResourceTargetTypeRoleTypeMap {
-    [WorkspaceInviteResourceType]: WorkspaceRoles
-  }
-}
-
-export type WorkspaceInviteResourceTarget = InviteResourceTarget<
-  typeof WorkspaceInviteResourceType,
-  WorkspaceRoles
->
+export { WorkspaceInviteResourceTarget } from '@/modules/workspacesCore/domain/types'
