+++ packages/server/modules/workspacesCore/domain/constants.ts
@@ -1,0 +1,1 @@
+export const WorkspaceInviteResourceType = 'workspace'
