+++ packages/server/modules/workspacesCore/domain/types.ts
@@ -1,4 +1,23 @@
+import { InviteResourceTarget } from '@/modules/serverinvites/domain/types'
+import { WorkspaceInviteResourceType } from '@/modules/workspacesCore/domain/constants'
 import { WorkspaceRoles } from '@speckle/shared'
+
+declare module '@/modules/serverinvites/domain/types' {
+  interface InviteResourceTargetTypeMap {
+    workspace: 'workspace'
+  }
+}
+
+declare module '@/modules/serverinvites/helpers/core' {
+  interface ResourceTargetTypeRoleTypeMap {
+    [WorkspaceInviteResourceType]: WorkspaceRoles
+  }
+}
+
+export type WorkspaceInviteResourceTarget = InviteResourceTarget<
+  typeof WorkspaceInviteResourceType,
+  WorkspaceRoles
+>
 
 export type Workspace = {
   id: string
